===Tier 3 Part 2=== -Added visualization of the player's upgrade tree -Added gameobjects (upgrade points) that when collected, unlock a specific path in the upgrade tree
(Will do a later commit to patch/expand/fix a few last minute details)

+++ Assets/ScriptableObjects/PlayerInfo.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using UnityEngine;
 /*
  * Written By: Gianni Coladonato
- * Date Created: 08-12-2025 | Last Modified: 26-04-2025
+ * Date Created: 08-12-2025 | Last Modified: 27-04-2025
  * 
  * This script is used to store player info between scenes
  */
@@ -20,5 +20,5 @@ public class PlayerInfo : ScriptableObject
     public bool keyB;
     public List<GameObject> weapons;
     public float atkBoost;
-    public int spdBoost;
+    public float spdBoost;
 }
